requirements.txt
numpy>=1.23.5,<2.0.0
pandas
scikit-learn
matplotlib
seaborn
boto3
PyYAML
tensorflow==2.16.1
streamlit
altair
wordcloud
palettable
nltk
selenium
beautifulsoup4
unidecode
joblib
pickle5